--- frontend/src/metabase/query_builder/components/filters/modals/FilterEditor/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./FilterEditor";
