+++ frontend/src/metabase/query_builder/components/filters/modals/FilterModal/FilterModal.tsx
@@ -1,10 +1,13 @@
-import React from "react";
+import React, { useState } from "react";
 import { t } from "ttag";
 import Question from "metabase-lib/lib/Question";
 import StructuredQuery from "metabase-lib/lib/queries/StructuredQuery";
 import Button from "metabase/core/components/Button";
 import ModalContent from "metabase/components/ModalContent";
-import FilterEditor from "metabase/query_builder/components/filters/modals/FilterEditor";
+import TabContent from "metabase/core/components/TabContent";
+import TabList from "metabase/core/components/TabList";
+import Tab from "metabase/core/components/Tab";
+import TabPanel from "metabase/core/components/TabPanel";
 
 export interface FilterModalProps {
   question: Question;
@@ -29,8 +32,32 @@ const FilterModal = ({
       ]}
       onClose={onClose}
     >
-      <FilterEditor query={query} />
+      <FilterModalBody query={query} />
     </ModalContent>
+  );
+};
+
+interface FilterModalBodyProps {
+  query: StructuredQuery;
+}
+
+const FilterModalBody = ({ query }: FilterModalBodyProps): JSX.Element => {
+  const [tab, setTab] = useState(0);
+  const sections = query.topLevelFilterFieldOptionSections();
+
+  return (
+    <TabContent value={tab} onChange={setTab}>
+      <TabList>
+        {sections.map((section, index) => (
+          <Tab key={index} value={index} icon={index > 0 ? "link" : undefined}>
+            {section.name}
+          </Tab>
+        ))}
+      </TabList>
+      {sections.map((section, index) => (
+        <TabPanel key={index} value={index} />
+      ))}
+    </TabContent>
   );
 };
